-- Fixed various warnings about conversions between unsigned/doubles
-- Commented out unusused stuff in Wassamon's driver codes; can probably
   go altogether but that's for Andrew to assess.

-- Added divergence and pore pressure to output of axisym poro-elastic
   traction elements.

-- Moved variable into PARANOID only block in triangle code.

-- Changed various atan to atan2 and initialised values in values
   in Wassamon's Subparametric_Telements.h.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@714 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ demo_drivers/fourier_decomposed_helmholtz/sphere_scattering/unstructured_sphere_scattering.cc
@@ -63,7 +63,7 @@ namespace PlanarWave
  unsigned N_terms=100;
 
  /// Wave number
- unsigned K=3.0*MathematicalConstants::Pi;
+ double K=3.0*MathematicalConstants::Pi;
 
  /// Imaginary unit 
  std::complex<double> I(0.0,1.0);
